RF: Gather interlinked utilities in distribution.utils (just moving)
